started backend

+++ mypro/src/main/java/com/mypro/mypro/repository/chatsRepository.java
@@ -3,5 +3,5 @@ package com.mypro.mypro.repository;
 import com.mypro.mypro.model.chats;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface chatsRepository extends JpaRepository<chats, Long> {
+public interface chatsRepository extends JpaRepository<chats, Integer> {
 }
