+++ mypro/src/main/java/com/mypro/mypro/repository/messagesRepository.java
@@ -3,5 +3,5 @@ package com.mypro.mypro.repository;
 import com.mypro.mypro.model.messages;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface messagesRepository extends JpaRepository<messages, Long> {
+public interface messagesRepository extends JpaRepository<messages, Integer> {
 }
