+++ mypro/src/main/java/com/mypro/mypro/repository/notificationsRepository.java
@@ -3,5 +3,5 @@ package com.mypro.mypro.repository;
 import com.mypro.mypro.model.notifications;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface notificationsRepository extends JpaRepository<notifications, Long> {
+public interface notificationsRepository extends JpaRepository<notifications, Integer> {
 }
