+++ mypro/src/main/java/com/mypro/mypro/repository/staffRepository.java
@@ -3,6 +3,6 @@ package com.mypro.mypro.repository;
 import com.mypro.mypro.model.staff;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface staffRepository extends JpaRepository<staff, Long> {
+public interface staffRepository extends JpaRepository<staff, Integer> {
 
 }
